refactor: delete old test file for announcement and split into three files for modularity
